requirements.txt
wheel
pika
Flask
flask-recaptcha
psycopg2-binary
uwsgi
chia-blockchain
geoip2